Add files via upload
🔧 Objectives & Overview
The goal of this project was to build a comprehensive data model and multi-page interactive dashboard in Power BI to analyse sales, forecast demand, and uncover business insights for a fictional pizzeria business. The key learning outcome was to model datasets using a star schema, build relationships, and leverage DAX for deeper analysis.